+++ android/src/main/java/com/facebook/flipper/android/FlipperConnectionImpl.java
@@ -40,9 +40,16 @@ class FlipperConnectionImpl implements FlipperConnection {
     sendArray(method, params);
   }
 
+  @Override
+  public void send(String method, String params) {
+    sendRaw(method, params);
+  }
+
   public native void sendObject(String method, FlipperObject params);
 
   public native void sendArray(String method, FlipperArray params);
+
+  public native void sendRaw(String method, String params);
 
   @Override
   public native void reportErrorWithMetadata(String reason, String stackTrace);