+++ desktop/plugins/public/navigation/plugin.tsx
@@ -189,6 +189,3 @@ export function plugin(client: PluginClient<Events, Methods>) {
     },
   };
 }
-
-/* @scarf-info: do not remove, more info: https://fburl.com/scarf */
-/* @scarf-generated: flipper-plugin index.js.template 0bfa32e5-fb15-4705-81f8-86260a1f3f8e */